+++ napari/layers/image_surface_mixin.py
