+++ napari/layers/surface/surface.py
@@ -6,7 +6,7 @@ from ...utils.colormaps import AVAILABLE_COLORMAPS
 from ...utils.event import Event
 from ..base import Layer
 from ..layer_utils import calc_data_range
-from ..mixins import ImageSurfaceMixin
+from ..image_surface_mixin import ImageSurfaceMixin
 
 
 # Mixin must come before Layer
